session-1-time-series-forecasting/requirements.txt
# Requirements for Session 1: Time Series Forecasting
azure-ai-ml>=1.12.0
azure-identity>=1.15.0
pandas>=2.0.0
numpy>=1.24.0
